splitCondor now takes output location argument
splitCondor now takes a compulsory argument specifying the location the output should be saved to. This is neccessary to allow the running of jobs on a local condor cluster, as logs are saved in the output directory, and the job merger needs the location of the files to be merged in the split file.